requirements.txt
# -----------------------------
# SERVER
# -----------------------------
fastapi==0.110.0
uvicorn[standard]

# -----------------------------
# DATABASE
# -----------------------------
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

# -----------------------------
# AUTHENTICATION
# -----------------------------
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
werkzeug==3.0.2
pillow>=10.1.0

# -----------------------------
# VALIDATION
# -----------------------------
pydantic>=2.9.2
email-validator==2.1.1
packaging==24.0

# -----------------------------
# FILE UPLOADS
# -----------------------------
python-multipart==0.0.9

# -----------------------------
# HTTP (FIXED COMPATIBILITY)
# -----------------------------
httpx>=0.28.1

# -----------------------------
# METRICS
# -----------------------------
prometheus-client==0.20.0

# -----------------------------
# FAST JSON
# -----------------------------
orjson==3.9.15

# -----------------------------
# PDF PARSING
# -----------------------------
pdfplumber==0.10.3
pdfminer.six==20221105
pypdf==4.1.0

# -----------------------------
# WORD DOCS
# -----------------------------
python-docx==1.1.0

# -----------------------------
# PDF GENERATION
# -----------------------------
reportlab==4.1.0

# -----------------------------
# DATA
# -----------------------------
numpy==1.26.4
pandas==2.2.1

# -----------------------------
# VECTOR SEARCH
# -----------------------------
faiss-cpu==1.8.0

# -----------------------------
# AI (FIXED 🔥)
# -----------------------------
openai>=1.30.0
google-genai==1.55.0

# -----------------------------
# IMAGE PROCESSING
# -----------------------------
albumentations==2.0.8

# -----------------------------
# ASYNC
# -----------------------------
anyio>=4.8.0

# -----------------------------
# STABILITY
# -----------------------------
typing-extensions>=4.9.0
setuptools==69.5.1